PostgreSQL database change commit
- database engine and structure changed
- generalization table person added for app_user and patient
- images are now stored in cloud

+++ backend/src/main/java/com/website/demo/webSocket/WebSocketMessage.java
@@ -7,7 +7,7 @@ public class WebSocketMessage {
 
     private String date;
     private String content;
-    private Long senderId;
-    private Long receiverId;
+    private Integer senderId;
+    private Integer receiverId;
 
 }
